Reuse the dynamic api prefix resolution for loading term.
Remove obsolete AsyncPublicViewActions.ts and their test file.
